Adds check that ref seqnames do not contain semicolon, which clashes with XA:Z
If a user's input reference sequences have a semicolon in the name,
this clashes with the use of the semicolon as a delimiter in the XA:Z
tag. It may be useful to warn users of this potential issue during the
bwa index step. For example, if users use xa2multi.pl to make multiple
entries from the XA tag and pass to samtools view, samtools view will
only look for the refname prior to the semicolon.