Update bwape.c
Addition of clipping score parameter, so users who are doing alignment with samples that may have legitimately unmapped mates (where the quality of the reference is known to be poor or possibly incomplete).  You can disable SW of course, or you can just discourage clipping which we've found to be the the primary error mode for the SW step.   Currently, bwa can generate some pretty strange clipped alignments, with only 22 base pairs matching out of, say, 100.  This option can be used to tune that behavior.