r930: use 3rd round seeding by default
This strategy is similar to the seeding heuristic used by LAST. When it is used
alone, it is not as accurate as the current seeding strategy at least for short
reads. However, it may do a better job for a long contig mapped to multiple ALT
contigs. This seeding strategy is also relatively cheap to perform.