Fixed typo in secondary alignment flag
When the -M flag is used (opt->flag & MEM_F_NO_MULTI == true), supplementary alignments are supposed to be marked as secondary (ie. read flag bit 0x100 should be set).  However, due to what appears to be a typo in the code, the flag bit 0x10000 is set instead.  This silently has no effect, as 0x10000 is not a valid SAM flag bit, and the net result is that the intended 0x100 bit remains unset.  Thus, when the -M flag is supplied, supplementary alignments are not correctly marked, and so are not filtered out in downstream steps.